SWEngineer/requirements.txt
langchain-chroma
langchain-openai
langchain-teddynote
langchain-community
langchain-experimental
streamlit
pymupdf
python-dotenv
chromadb
pymongo